be/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
django-cors-headers==4.3.1
django-filter==23.5
django-celery-beat==2.5.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
python-dateutil==2.8.2
drf-spectacular==0.26.5
